refactor(server): exceptions shared between field relation and object metadata names
