+++ packages/twenty-server/src/engine/metadata-modules/relation-metadata/relation-metadata.service.ts
@@ -19,7 +19,6 @@ import {
   RelationMetadataException,
   RelationMetadataExceptionCode,
 } from 'src/engine/metadata-modules/relation-metadata/relation-metadata.exception';
-import { InvalidStringException } from 'src/engine/metadata-modules/utils/exceptions/invalid-string.exception';
 import { validateFieldNameAvailabilityOrThrow } from 'src/engine/metadata-modules/utils/validate-field-name-availability.utils';
 import { WorkspaceMetadataVersionService } from 'src/engine/metadata-modules/workspace-metadata-version/services/workspace-metadata-version.service';
 import { generateMigrationName } from 'src/engine/metadata-modules/workspace-migration/utils/generate-migration-name.util';
@@ -33,8 +32,8 @@ import { computeObjectTargetTable } from 'src/engine/utils/compute-object-target
 import { WorkspaceCacheStorageService } from 'src/engine/workspace-cache-storage/workspace-cache-storage.service';
 import { WorkspaceMigrationRunnerService } from 'src/engine/workspace-manager/workspace-migration-runner/workspace-migration-runner.service';
 import { BASE_OBJECT_STANDARD_FIELD_IDS } from 'src/engine/workspace-manager/workspace-sync-metadata/constants/standard-field-ids';
-import { validateObjectMetadataInputNameOrThrow } from 'src/engine/metadata-modules/object-metadata/utils/validate-object-metadata-input.util';
 
+import { validateMetadataNameOrThrow } from 'src/engine/metadata-modules/utils/validate-metadata-name.utils';
 import {
   RelationMetadataEntity,
   RelationMetadataType,
@@ -67,17 +66,13 @@ export class RelationMetadataService extends TypeOrmQueryService<RelationMetadat
     );
 
     try {
-      validateObjectMetadataInputNameOrThrow(relationMetadataInput.fromName);
-      validateObjectMetadataInputNameOrThrow(relationMetadataInput.toName);
+      validateMetadataNameOrThrow(relationMetadataInput.fromName);
+      validateMetadataNameOrThrow(relationMetadataInput.toName);
     } catch (error) {
-      if (error instanceof InvalidStringException) {
-        throw new RelationMetadataException(
-          `Characters used in name "${relationMetadataInput.fromName}" or "${relationMetadataInput.toName}" are not supported`,
-          RelationMetadataExceptionCode.INVALID_RELATION_INPUT,
-        );
-      } else {
-        throw error;
-      }
+      throw new RelationMetadataException(
+        error.message,
+        RelationMetadataExceptionCode.INVALID_RELATION_INPUT,
+      );
     }
 
     await this.validateCreateRelationMetadataInput(
