+++ packages/twenty-server/src/engine/metadata-modules/utils/exceptions/name-is-reserved-keyword.exception.ts
@@ -1,0 +1,7 @@
+export class NameIsReservedKeywordException extends Error {
+  constructor(name: string) {
+    const message = `The name "${name}" is not available`;
+
+    super(message);
+  }
+}
