+++ packages/twenty-server/src/engine/metadata-modules/utils/exceptions/name-not-camel-case.exception.ts
@@ -1,0 +1,7 @@
+export class NameIsNotInCamelCase extends Error {
+  constructor(name: string) {
+    const message = `Name should be in camelCase: ${name}`;
+
+    super(message);
+  }
+}
