+++ packages/twenty-server/src/engine/metadata-modules/utils/exceptions/name-too-short.exception.ts
@@ -1,0 +1,7 @@
+export class NameTooShortException extends Error {
+  constructor(name: string) {
+    const message = `Input is too short: ${name}`;
+
+    super(message);
+  }
+}
