+++ packages/twenty-server/src/engine/metadata-modules/utils/validate-metadata-name-is-camel-case.utils.ts
@@ -1,0 +1,8 @@
+import camelCase from 'lodash.camelcase';
+import { NameIsNotInCamelCase } from 'src/engine/metadata-modules/utils/exceptions/name-not-camel-case.exception';
+
+export const validateMetadataNameIsCamelCaseOrThrow = (name: string) => {
+  if (name !== camelCase(name)) {
+    throw new NameIsNotInCamelCase(name);
+  }
+};
