+++ packages/twenty-server/src/engine/metadata-modules/utils/validate-metadata-name-is-not-too-long.utils copy.ts
@@ -1,0 +1,8 @@
+import { NameTooLongException } from 'src/engine/metadata-modules/utils/exceptions/name-too-long.exception';
+import { exceedsDatabaseIdentifierMaximumLength } from 'src/engine/metadata-modules/utils/validate-database-identifier-length.utils';
+
+export const validateMetadataNameIsNotTooLongOrThrow = (name: string) => {
+  if (exceedsDatabaseIdentifierMaximumLength(name)) {
+    throw new NameTooLongException(name);
+  }
+};
